+++ CHANGELOG.md
@@ -10,6 +10,7 @@ adheres to the format set out by [Keep a Changelog](https://keepachangelog.com/e
 - [`Breaking change`] Make SignedTransaction.Hash() output string to be consistent with other representations
 - Add more functions available from under util
 - Add CoinBatchTransferPayload for sending multiple amounts to multiple addresses
+- [`Fix`] Block APIs will now pull the rest of the transactions for the block automatically
 
 # v0.2.0 (6/10/2024)